feat: Move a couple files to ts while keeping the project running
