+++ sources/canvas/preview-animation.js
@@ -1,8 +1,8 @@
 import { previewCanvas, previewCtx } from "./preview-canvas.js";
 import { state } from "../state/state.js";
-import { FRAME_SIZE, ANIMATION_CONFIGS } from "../state/constants.js";
+import { FRAME_SIZE, ANIMATION_CONFIGS } from "../state/constants.ts";
 import { get2DContext, drawTransparencyBackground } from "./canvas-utils.js";
-import { applyTransparencyMaskToCanvas } from "./mask.js";
+import { applyTransparencyMaskToCanvas } from "./mask.ts";
 import { canvas } from "./renderer.js";
 import { customAnimations } from "../custom-animations.js";
