+++ sources/components/tree/ItemWithRecolors.js
@@ -4,7 +4,7 @@ import { state, getSelectionGroup, selectItem } from "../../state/state.js";
 import { drawRecolorPreview } from "../../canvas/palette-recolor.js";
 import { getPaletteOptions } from "../../state/palettes.js";
 import { PaletteSelectModal } from "./PaletteSelectModal.js";
-import { COMPACT_FRAME_SIZE, FRAME_SIZE } from "../../state/constants.js";
+import { COMPACT_FRAME_SIZE, FRAME_SIZE } from "../../state/constants.ts";
 
 export const ItemWithRecolors = {
   view: function (vnode) {
