+++ sources/components/tree/ItemWithVariants.js
@@ -2,7 +2,7 @@
 import classNames from "classnames";
 import { state, getSelectionGroup, selectItem } from "../../state/state.js";
 import { getLayersToLoad } from "../../state/meta.js";
-import { COMPACT_FRAME_SIZE, FRAME_SIZE } from "../../state/constants.js";
+import { COMPACT_FRAME_SIZE, FRAME_SIZE } from "../../state/constants.ts";
 import { capitalize } from "../../utils/helpers.js";
 
 export const ItemWithVariants = {
