+++ sources/utils/zip-helpers.js
@@ -3,7 +3,7 @@ import {
   FRAME_SIZE,
   STANDARD_ANIMATION_FRAMES_PER_ROW,
   DIRECTIONS,
-} from "../state/constants.js";
+} from "../state/constants.ts";
 import { drawFramesToCustomAnimation } from "../canvas/draw-frames.js";
 import { customAnimationSize } from "../custom-animations.js";
 import {
