+++ tests/node/scripts/generateSources/items_spec.js
@@ -3,7 +3,7 @@ import assert from "node:assert/strict";
 import fs from "node:fs";
 import os from "node:os";
 import path from "node:path";
-import { ANIMATION_DEFAULTS } from "../../../../sources/state/constants.js";
+import { ANIMATION_DEFAULTS } from "../../../../sources/state/constants.ts";
 import {
   parseItem,
   getRequiredSexes,
