+++ tests/state/zip_spec.js
@@ -23,7 +23,7 @@ import {
 } from "../../sources/state/zip.js";
 import { resetState } from "../../sources/state/hash.js";
 import { state } from "../../sources/state/state.js";
-import { ANIMATIONS, DIRECTIONS } from "../../sources/state/constants.js";
+import { ANIMATIONS, DIRECTIONS } from "../../sources/state/constants.ts";
 import { createFakeJSZip } from "../helpers/fake-jszip.js";
 
 /**
